+++ lib/cgraph/cgraph.h
@@ -564,12 +564,6 @@ CGRAPH_API int agobjkind(void *obj);
  * All uses of cgraph strings need to be freed using @ref agstrfree
  * in order to correctly maintain the reference count.
  *
- * @ref agcanonStr returns a pointer to a version of the input string
- * canonicalized for output for later re-parsing.
- * This includes quoting special characters and keywords.
- * It uses its own internal buffer, so the value will be lost on
- * the next call to @ref agcanonStr.
- *
  * @{
  */
 CGRAPH_API char *agstrdup(Agraph_t *, const char *);
@@ -605,7 +599,6 @@ CGRAPH_API char *agstrbind_html(Agraph_t *g, const char *);
 CGRAPH_API int agstrfree(Agraph_t *, const char *, bool is_html);
 ///< @param is_html Is the string being freed an HTML-like string?
 CGRAPH_API char *agstrcanon(char *, char *);
-CGRAPH_API char *agcanonStr(char *str); /* manages its own buf */
 /// @}
 
 /** @defgroup cgraph_attr attributes